% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite.R
\name{txregCollections}
\alias{txregCollections}
\title{list collections in AWS mongo server for txregnet}
\usage{
txregCollections(ignore = NULL, url = URL_txregInAWS(),
  db = "txregnet", cliparms = "--quiet --eval")
}
\arguments{
\item{ignore}{defaults to NULL, otherwise an integer vector telling which lines of mongo db.getCollectionNames() result should be ignored}

\item{url}{a valid mongodb URL}

\item{db}{character(1) db name}

\item{cliparms}{character(1) added parameters for mongo CLI call}
}
\value{
a character vector of collection names
}
\description{
list collections in AWS mongo server for txregnet
}
\examples{
if (verifyHasMongoCmd()) txregCollections()[seq_len(5)]
}
